+++ Automate/docs/release-notes.md
@@ -1,6 +1,9 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* Fixed some machines missing in Generic Mod Config Menu UI.
+
 ## 2.0.4
 Released 04 April 2024 for SMAPI 4.0.0 or later.